reuse Cache struct in extreme and string_agg
